[UPD] Corrección SignIn SignUp

+++ src/components/signin-signup/signIn-signUp.jsx
@@ -32,16 +32,42 @@ class SignInSignup extends Component {
         const { email, password, displayName } = this.state;
 
         try {
-            const { user } = await auth.createUserWithEmailAndPassword(email, password);
-            const userRef = await createUserProfile(user, { displayName });
+            const response = await auth.createUserWithEmailAndPassword(email, password)
+                .catch(error => {
+                    var errorCode = error.code;
+                    var errorMessage = error.message;
+                    if (errorCode == 'auth/weak-password') {
+                        Swal.fire({
+                            position: 'center',
+                            icon: 'warning',
+                            title: 'The password is too weak.',
+                            showConfirmButton: false,
+                            timer: 2000
+                        });
+                    } else {
+                        Swal.fire({
+                            position: 'center',
+                            icon: 'error',
+                            title: errorMessage,
+                            showConfirmButton: false,
+                            timer: 2000
+                        });
+                    }
+                    console.log(`Error on SignInSignup::handleSignUp -> ${error}`);
+                });
 
-            await Swal.fire({
-                position: 'center',
-                icon: 'success',
-                title: `Usuario ${displayName} ha sido Registrado`,
-                showConfirmButton: false,
-                timer: 2000
-            });
+            if (response != undefined) {
+                const { user } = response;
+                const userRef = await createUserProfile(user, { displayName });
+
+                await Swal.fire({
+                    position: 'center',
+                    icon: 'success',
+                    title: `Usuario ${displayName} ha sido Registrado`,
+                    showConfirmButton: false,
+                    timer: 2000
+                });
+            }
         } catch (error) {
             console.error(`Error on SignInSignup::handleSignUp -> ${error}`);
         }
@@ -51,18 +77,31 @@ class SignInSignup extends Component {
         const { email, password } = this.state;
 
         try {
-            const { user } = await auth.signInWithEmailAndPassword(email, password);
-            localStorage.setItem('userLoggedIn', user.email);
+            const response = await auth.signInWithEmailAndPassword(email, password)
+                .catch(error => {
+                    Swal.fire({
+                        position: 'center',
+                        icon: 'error',
+                        title: `Invalid Email and/or Password!`,
+                        showConfirmButton: false,
+                        timer: 2000
+                    });
+                });
 
-            await Swal.fire({
-                position: 'center',
-                icon: 'success',
-                title: `Bienvenid@ ${user.email} !`,
-                showConfirmButton: false,
-                timer: 2000
-            });
+            if (response !== undefined) {
+                const { user } = response;
+                localStorage.setItem('userLoggedIn', user.email);
 
-            this.props.history.push('/home');
+                await Swal.fire({
+                    position: 'center',
+                    icon: 'success',
+                    title: `Bienvenid@ ${user.email} !`,
+                    showConfirmButton: false,
+                    timer: 2000
+                });
+
+                this.props.history.push('/home');
+            }
         } catch (error) {
             console.error(`Error on SignInSignup::handleSignIn -> ${error}`);
         }
